[Core][Shop] Move emails from core to shop
