+++ src/Sylius/Bundle/ShopBundle/Resources/config/app/config.yml
@@ -2,6 +2,7 @@
 # (c) Paweł Jędrzejewski
 
 imports:
+    - { resource: "@SyliusShopBundle/Resources/config/app/sylius/sylius_mailer.yml" }
     - { resource: "@SyliusShopBundle/Resources/config/grids/account/order.yml" }
     - { resource: "@SyliusShopBundle/Resources/config/grids/product.yml" }
