+++ src/Sylius/Bundle/ShopBundle/Resources/config/app/sylius/sylius_mailer.yml
@@ -1,0 +1,26 @@
+# This file is part of the Sylius package.
+# (c) Paweł Jędrzejewski
+
+sylius_mailer:
+    emails:
+        contact_request:
+            subject: sylius.emails.contact_request.subject
+            template: "SyliusShopBundle:Email:contactRequest.html.twig"
+        order_confirmation:
+            subject: sylius.emails.order_confirmation.subject
+            template: "SyliusShopBundle:Email:orderConfirmation.html.twig"
+        user_registration:
+            subject: sylius.emails.user_registration.subject
+            template: "SyliusShopBundle:Email:userRegistration.html.twig"
+        password_reset:
+            subject: sylius.emails.user.password_reset.subject
+            template: "SyliusShopBundle:Email:passwordReset.html.twig"
+        reset_password_token:
+            subject: sylius.emails.user.password_reset.subject
+            template: "SyliusShopBundle:Email:passwordReset.html.twig"
+        reset_password_pin:
+            subject: sylius.emails.user.password_reset.subject
+            template: "SyliusShopBundle:Email:passwordReset.html.twig"
+        verification_token:
+            subject: sylius.emails.user.verification_token.subject
+            template: "SyliusShopBundle:Email:verification.html.twig"
