+++ src/Sylius/Bundle/ShopBundle/Resources/config/services.xml
@@ -14,7 +14,8 @@
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <imports>
         <import resource="services/controller.xml" />
-        <import resource="services/menu.xml" />
+        <import resource="services/email.xml" />
         <import resource="services/listeners.xml" />
+        <import resource="services/menu.xml" />
     </imports>
 </container>
