+++ src/Sylius/Bundle/ShopBundle/Resources/config/services/email.xml
@@ -13,7 +13,7 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="sylius.listener.user_mailer_listener" class="Sylius\Bundle\CoreBundle\EventListener\UserMailerListener">
+        <service id="sylius.listener.user_mailer_listener" class="Sylius\Bundle\ShopBundle\EventListener\UserMailerListener">
             <argument type="service" id="sylius.email_sender" />
             <argument type="service" id="sylius.context.channel" />
             <tag name="kernel.event_listener" event="sylius.user.password_reset.request.pin" method="sendResetPasswordPinEmail" />
@@ -22,15 +22,11 @@
             <tag name="kernel.event_listener" event="sylius.customer.post_register" method="sendUserRegistrationEmail" />
         </service>
 
-        <service id="sylius.email_manager.contact" class="Sylius\Bundle\CoreBundle\EmailManager\ContactEmailManager">
+        <service id="sylius.email_manager.contact" class="Sylius\Bundle\ShopBundle\EmailManager\ContactEmailManager">
             <argument type="service" id="sylius.email_sender" />
         </service>
 
-        <service id="sylius.email_manager.order" class="Sylius\Bundle\CoreBundle\EmailManager\OrderEmailManager">
-            <argument type="service" id="sylius.email_sender" />
-        </service>
-
-        <service id="sylius.email_manager.shipment" class="Sylius\Bundle\CoreBundle\EmailManager\ShipmentEmailManager">
+        <service id="sylius.email_manager.order" class="Sylius\Bundle\ShopBundle\EmailManager\OrderEmailManager">
             <argument type="service" id="sylius.email_sender" />
         </service>
     </services>
