+++ src/Sylius/Bundle/ShopBundle/Resources/views/Email/userRegistration.html.twig
