+++ src/Sylius/Bundle/ShopBundle/spec/EventListener/UserMailerListenerSpec.php
@@ -9,20 +9,18 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\CoreBundle\EventListener;
+namespace spec\Sylius\Bundle\ShopBundle\EventListener;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Sylius\Bundle\CoreBundle\EventListener\UserMailerListener;
 use Sylius\Bundle\CoreBundle\Mailer\Emails;
+use Sylius\Bundle\ShopBundle\EventListener\UserMailerListener;
 use Sylius\Bundle\UserBundle\EventListener\MailerListener;
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Channel\Model\ChannelInterface;
 use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\ShopUserInterface;
 use Sylius\Component\Mailer\Sender\SenderInterface;
-use Sylius\Component\Resource\Exception\UnexpectedTypeException;
-use Sylius\Component\User\Model\UserInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
 
 /**
